% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau.R
\name{kendall_tau}
\alias{kendall_tau}
\title{Calculating Kendall Tau using merge sort}
\usage{
kendall_tau(X)
}
\arguments{
\item{X}{A matrix with two columns of real numbers with no ties within the same column}
}
\value{
\code{tau}       : Kendall Tau correlation between the two columns
}
\description{
Calculating Kendall Tau using merge sort
}
\examples{
# 5 x 2 matrix
X <- matrix(c(4,5,3,1,2,6,8,9,7,10),5,2)
# Find Kendall tau
kendall_tau(X)
}
