% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mergeSort}
\alias{mergeSort}
\title{Finding inversion count of a vector using merge sort by divide-and-conquer}
\usage{
mergeSort(x, l, r)
}
\arguments{
\item{x}{A vector containing real numbers with no ties}

\item{l}{Index of the beginning (inclusive) of the subset to be sorted}

\item{r}{Index of the end (exclusive) of the subset to be sorted}
}
\value{
\code{inv} Inversion count
}
\description{
Finding inversion count of a vector using merge sort by divide-and-conquer
}
