\name{PrecipGL}
\alias{PrecipGL}
\docType{data}
\title{Annual precipitation, inches, Great Lakes, 1900-1986}
\description{
  Annual precipitation, 1900-1986, Entire Great Lakes
}

\usage{data(PrecipGL)}

\format{
  The format is:
 Time-Series [1:87] from 1900 to 1986: 31.7 29.8 31.7 33.1 31.3 ...
}

\source{
Appendix A of H. Roberts (1992) 
"Data Analysis for Managers" published by Scientific Press.

Original Source: Great Lake Water Levels, U.S. Dept of Commerce, Rockville MD          
U.S. Lake Survey, Detroit, MI, US Army Corps of Engineers             
}
\examples{
data(PrecipGL)
plot(PrecipGL)
}
\keyword{datasets}


