\name{kendall-package}
\alias{kendall-package}
\alias{kendall}
\docType{package}
\title{Kendall Tau correlation coefficient}
\description{
  Calculating Kendall Tau using merge sort. To learn more about kendall, start with the vignettes: browseVignettes("kendall")
}
\details{More details about kendall can be found in the readme: https://github.com/lapsumchan/kendall/blob/master/README.md
}
\author{
Lap Sum Chan

Maintainer: Lap Sum Chan <lapsum@umich.edu>
}
