\name{kendall-package}
\alias{kendall-package}
\alias{kendall}
\docType{package}
\title{Kendall Tau correlation coefficient}
\description{
  Calculating Kendall Tau using merge sort.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Lap Sum Chan

Maintainer: Lap Sum Chan <lapsum@umich.edu>
}
\examples{}
